start_screen_di more fix
