+++ featureStartScreen/src/main/java/com/bestscore/featurestartscreen/StartScreenFragment.kt
@@ -1,6 +1,7 @@
 package com.bestscore.featurestartscreen
 
 import android.content.Context
+import android.graphics.Canvas
 import android.os.Bundle
 import android.view.View
 import androidx.core.view.isVisible
@@ -11,12 +12,17 @@ import androidx.lifecycle.get
 import androidx.lifecycle.lifecycleScope
 import androidx.recyclerview.widget.ItemTouchHelper
 import androidx.recyclerview.widget.LinearLayoutManager
+import androidx.recyclerview.widget.RecyclerView
+import androidx.recyclerview.widget.RecyclerView.ItemDecoration
 import by.kirich1409.viewbindingdelegate.viewBinding
 import com.bestscore.core.navigation.navigate
 import com.bestscore.featurestartscreen.databinding.FragmentStartScreenBinding
 import com.bestscore.featurestartscreen.di.StartScreenComponentViewModel
+import com.bestscore.featurestartscreen.swipe.SwipeController
+import com.bestscore.featurestartscreen.swipe.SwipeControllerActions
 import com.bestscore.utils.makeToast
 import javax.inject.Inject
+
 
 class StartScreenFragment : Fragment(R.layout.fragment_start_screen) {
 
@@ -40,9 +46,8 @@ class StartScreenFragment : Fragment(R.layout.fragment_start_screen) {
         )
     }
 
-    private val swipeController:SwipeController = SwipeController()
-
-    private val itemTouchHelper:ItemTouchHelper = ItemTouchHelper(swipeController)
+    //private val swipeController: SwipeController = SwipeController()
+   // private val itemTouchHelper: ItemTouchHelper = ItemTouchHelper(swipeController)
 
     override fun onAttach(context: Context) {
         ViewModelProvider(this)
@@ -79,10 +84,26 @@ class StartScreenFragment : Fragment(R.layout.fragment_start_screen) {
         binding.rvLastTemplates.apply {
             adapter = this@StartScreenFragment.adapter
             layoutManager = LinearLayoutManager(requireContext())
+
+            var swipeController = SwipeController(object : SwipeControllerActions() {
+                override fun onRightClicked(position: Int) {
+                    makeToast("Click click")
+                }
+            })
+            val itemTouchHelper = ItemTouchHelper(swipeController)
+
+
             itemTouchHelper.attachToRecyclerView(this)
+
+            this.addItemDecoration(object : ItemDecoration() {
+                override fun onDraw(c: Canvas, parent: RecyclerView, state: RecyclerView.State) {
+                    swipeController.onDraw(c)
+                }
+            })
         }
 
     }
+
 
     private fun collectFlow() {
         viewLifecycleOwner.lifecycleScope.launchWhenStarted {
